+++ babel.config.js
@@ -1,6 +1,28 @@
-module.exports = function (api) {
+module.exports = (api) => {
   api.cache(true)
+
   return {
     presets: ['babel-preset-expo'],
+    plugins: [
+      [
+        'module-resolver',
+        {
+          extensions: [
+            '.ts',
+            '.tsx',
+            '.android.ts',
+            '.android.tsx',
+            '.ios.ts',
+            '.ios.tsx',
+          ],
+          root: ['./src'],
+          alias: {
+            components: './src/components',
+            screens: './src/screens',
+            libs: './src/libs',
+          },
+        },
+      ],
+    ],
   }
 }
