+++ src/components/WineTitle/styles.ts
@@ -1,0 +1,7 @@
+import { StyleSheet } from 'react-native'
+
+const styles = StyleSheet.create({
+  wineName: { fontSize: 20, fontWeight: 'bold', display: 'flex' },
+})
+
+export default styles
