+++ src/libs/helpers/isAllowedPassword.ts
@@ -1,0 +1,17 @@
+/** @function
+ * @name isAllowedPassword
+ * Returns true if string is a password which passes all the tests.
+ *
+ * An allowed password must contain at least 8 symbols, at least one uppercase letter or at least 1 numeric
+ *
+ * @param {string} password - string to check
+ * */
+function isAllowedPassword(password: string): boolean {
+  const passwordRegexp = /(?:.*[^А-Яа-яЁё]*)(?=^.{8,}$)(?=.*[a-z])((?=.*[0-9])|(?=.*[A-Z]))/
+
+  if (!password) return false
+
+  return password.search(passwordRegexp) === 0
+}
+
+export default isAllowedPassword
