+++ src/screens/Auth/components/UserTypeSwitcher/styles.ts
@@ -1,0 +1,14 @@
+import { StyleSheet } from 'react-native'
+
+const styles = StyleSheet.create({
+  container: {
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'space-between',
+    flexDirection: 'row',
+    position: 'absolute',
+    top: '15%',
+  },
+})
+
+export default styles
