+++ src/screens/Auth/screens/ConfirmAge/styles.ts
@@ -1,0 +1,7 @@
+import { StyleSheet } from 'react-native'
+
+const styles = StyleSheet.create({
+  container: {},
+})
+
+export default styles
