+++ src/screens/Auth/screens/SignIn/index.tsx
@@ -1,15 +1,24 @@
-import React, { ReactElement, useState, useEffect } from 'react'
-import { StyleSheet, Text, View, TouchableOpacity, Linking } from 'react-native'
+import React, { useState, useEffect } from 'react'
+import { Text, View, TouchableOpacity, Linking } from 'react-native'
+import { StackScreenProps } from '@react-navigation/stack'
 import {
   useFonts,
   Merriweather_400Regular,
   Merriweather_700Bold,
 } from '@expo-google-fonts/merriweather'
 import { AppLoading } from 'expo'
-import LabeledInput from '../../molecules/LabeledInput'
-import { isEmail, isAllowedPassword } from '../../helpers'
 
-function SignInScreen(): ReactElement {
+import { isEmail, isAllowedPassword } from 'libs/helpers'
+import LabeledInput from 'components/TextInput'
+
+import { TParamList } from 'screens/Auth/types'
+
+import styles from './styles'
+
+export type TOuterProps = StackScreenProps<TParamList, 'Login'>
+type TProps = TOuterProps
+
+const SignInScreen: React.FC<TProps> = () => {
   const [userEmail, getUserEmail] = useState('')
   const [userPassword, getUserPassword] = useState('')
 
@@ -72,40 +81,3 @@ function SignInScreen(): ReactElement {
 }
 
 export default SignInScreen
-
-const styles = StyleSheet.create({
-  header: {
-    fontSize: 30,
-    fontFamily: 'Merriweather_700Bold',
-    color: '#C23232',
-  },
-  helper: {
-    paddingTop: 38,
-    paddingLeft: 20,
-    alignSelf: 'flex-start',
-    fontSize: 16,
-    color: '#4A7DFF',
-    fontFamily: 'Merriweather_400Regular',
-  },
-  container: {
-    paddingTop: 136,
-    flex: 1,
-    alignItems: 'center',
-    backgroundColor: 'white',
-  },
-  buttonStyle: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-    marginTop: 40,
-    width: 320,
-    maxHeight: 56,
-    backgroundColor: '#C23232',
-    borderRadius: 10,
-  },
-  buttonText: {
-    fontSize: 20,
-    color: '#fff',
-    fontFamily: 'Merriweather_400Regular',
-  },
-})
