+++ src/screens/Auth/screens/SignIn/styles.ts
@@ -1,0 +1,40 @@
+import { StyleSheet } from 'react-native'
+
+const styles = StyleSheet.create({
+  header: {
+    fontSize: 30,
+    fontFamily: 'Merriweather_700Bold',
+    color: '#C23232',
+  },
+  helper: {
+    paddingTop: 38,
+    paddingLeft: 20,
+    alignSelf: 'flex-start',
+    fontSize: 16,
+    color: '#4A7DFF',
+    fontFamily: 'Merriweather_400Regular',
+  },
+  container: {
+    paddingTop: 136,
+    flex: 1,
+    alignItems: 'center',
+    backgroundColor: 'white',
+  },
+  buttonStyle: {
+    flex: 1,
+    alignItems: 'center',
+    justifyContent: 'center',
+    marginTop: 40,
+    width: 320,
+    maxHeight: 56,
+    backgroundColor: '#C23232',
+    borderRadius: 10,
+  },
+  buttonText: {
+    fontSize: 20,
+    color: '#fff',
+    fontFamily: 'Merriweather_400Regular',
+  },
+})
+
+export default styles
