+++ src/screens/Auth/screens/SignUp/styles.ts
@@ -1,0 +1,32 @@
+import { StyleSheet } from 'react-native'
+
+const styles = StyleSheet.create({
+  header: {
+    fontSize: 30,
+    fontFamily: 'Merriweather_700Bold',
+    color: '#C23232',
+  },
+  container: {
+    paddingTop: 136,
+    flex: 1,
+    alignItems: 'center',
+    backgroundColor: 'white',
+  },
+  buttonStyle: {
+    flex: 1,
+    alignItems: 'center',
+    justifyContent: 'center',
+    marginTop: 60,
+    width: 320,
+    maxHeight: 56,
+    backgroundColor: '#C23232',
+    borderRadius: 10,
+  },
+  buttonText: {
+    fontSize: 20,
+    color: '#fff',
+    fontFamily: 'Merriweather_400Regular',
+  },
+})
+
+export default styles
