+++ src/screens/Auth/types.ts
@@ -1,0 +1,5 @@
+export type TParamList = {
+  ConfirmAge: undefined
+  Login: undefined
+  Register: undefined
+}
