+++ tsconfig.json
@@ -7,6 +7,12 @@
     "noEmit": true,
     "skipLibCheck": true,
     "resolveJsonModule": true,
-    "strict": true
+    "strict": true,
+    "baseUrl": "./",
+    "paths": {
+      "components": ["./src/components"],
+      "screens": ["./src/screens"],
+      "libs": ["./src/libs"]
+    }
   }
 }
